Use net/http
This commit migrates this example to use Go's built-in net/http package,
which is the standard way to build servers in Go. This new example just
renders a simple static page, like the default express-hello-world
example.

This new version doesn't use a Dockerfile, since it demonstrates
Render's native support for Go binaries, but I'm open to the argument
that showing a Dockerfile may be more instructive in other ways.